A lot of cleanups.
(Logical change 1.2481)

+++ include/aal/bitops.h
@@ -6,14 +6,18 @@
 #ifndef BITOPS_H
 #define BITOPS_H
 
-#define _ROUND_UP(x,n) (((x)+(n)-1u) & ~((n)-1u))
-#define ROUND_UP(x) _ROUND_UP(x,8LL)
+#define _ROUND_UP(x, n) (((x)+(n) - 1u) & ~((n) - 1u))
+#define ROUND_UP(x) _ROUND_UP(x, 8ll)
 
-extern inline int aal_set_bit (int nr, void *addr);
-extern inline int aal_clear_bit (int nr, void *addr);
-extern inline int aal_test_bit(int nr, const void *addr);
-extern inline int aal_find_first_zero_bit (const void *vaddr, unsigned size);
-extern inline int aal_find_next_zero_bit (const void *vaddr, unsigned size, unsigned offset);
+extern inline int aal_set_bit (unsigned long nr, void *addr);
+extern inline int aal_clear_bit (unsigned long nr, void *addr);
+extern inline int aal_test_bit(unsigned long nr, const void *addr);
+
+extern inline int aal_find_first_zero_bit (const void *vaddr, 
+    unsigned long size);
+
+extern inline int aal_find_next_zero_bit (const void *vaddr, 
+    unsigned long size, unsigned long offset);
 
 #endif
